added in security check for uninstall.php
git-svn-id: https://plugins.svn.wordpress.org/custom-stock-widget/trunk@1109960 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,4 +1,8 @@
 <?php 
+if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	echo "ACCESS FORBIDDEN";
+    exit();
+}
 global $wpdb;
 $wpdb->query( "DROP TABLE IF EXISTS " . $wpdb->prefix . "stock_widgets" );
 delete_option('stock_widget_per_category_stock_lists');
